Move things out to declutter lib.rs
